Update sather to 1.2.3.
Also update the tk dep to no longer demand tk83. It does not seem to
work all that well with tk84, but it works about the same as it does
with tk83. And depend explicitly on tcl; that was missing.

Primary changes since 1.2.1 seem to be GPLv2 -> GPLv3 and
whitespace/reindenting, but there seems to be a bit more in there
too. There's also now a 12 meg kdevelop blob in the distfile, yay.

Changelog:

    GNU Sather 1.2.3 - Jul 7, 2007 - Michael R. Taylor

    * Made INT 32-bit even on 64-bit systems
    * Changed license to (GPLv3/LGPLv3) or later

There's no changelog for 1.2.2, dunno why.