Don't include <sys/types.h> until after config.h has had a chance to determine which features will be exposed.  Fixes build on Linux.
<sys/types.h> is included immediately after config.h in the distribution code,
suggesting that the purged hunk in the patch is now obsolete.

The history of patch-aa reveals that the <sys/types.h> include was the very
original purpose of this patch, and predates the Solaris portion of the patch.
The Solaris hunks don't look like they would benefit from including
<sys/types.h>.

No PKGREVISION bump, this change should only affect platforms where the
package failed to build.

Closes PR pkg/36404 with the fix suggested by Ondrej Tuma (thanks!)